Move NWB writer into save_poses.py registry
